style: fit sphinx-gallery style
